move block into version specific module
