+++ crates/papyrus_gateway/src/api/v0_3_0/v0_3_0_impl.rs
@@ -19,15 +19,16 @@ use super::{
     JsonRpcV0_3_0Server,
 };
 use crate::api::{BlockHashOrNumber, ContinuationToken, JsonRpcError, JsonRpcServerImpl};
-use crate::block::Block;
+use crate::block::get_block_header_by_number;
 use crate::state::StateUpdate;
 use crate::transaction::{
     Event, TransactionOutput, TransactionReceipt, TransactionReceiptWithStatus,
     TransactionWithType, Transactions,
 };
+use crate::v0_3_0::block::{Block, BlockHeader};
 use crate::{
-    get_block_header_by_number, get_block_number, get_block_status, get_block_txs_by_number,
-    get_latest_block_number, internal_server_error, ContinuationTokenAsStruct,
+    get_block_number, get_block_status, get_block_txs_by_number, get_latest_block_number,
+    internal_server_error, ContinuationTokenAsStruct,
 };
 
 /// Rpc server.
@@ -51,7 +52,7 @@ impl JsonRpcV0_3_0Server for JsonRpcServerV0_3_0Impl {
         let txn = self.storage_reader.begin_ro_txn().map_err(internal_server_error)?;
         let block_number = get_latest_block_number(&txn)?
             .ok_or_else(|| ErrorObjectOwned::from(JsonRpcError::NoBlocks))?;
-        let header = get_block_header_by_number(&txn, block_number)?;
+        let header: BlockHeader = get_block_header_by_number(&txn, block_number)?;
 
         Ok(BlockHashAndNumber { block_hash: header.block_hash, block_number })
     }
@@ -161,7 +162,7 @@ impl JsonRpcV0_3_0Server for JsonRpcServerV0_3_0Impl {
 
         // Get the block header for the block hash and state root.
         let block_number = get_block_number(&txn, block_id)?;
-        let header = get_block_header_by_number(&txn, block_number)?;
+        let header: BlockHeader = get_block_header_by_number(&txn, block_number)?;
 
         // Get the old root.
         let parent_block_number = get_block_number(
@@ -171,7 +172,8 @@ impl JsonRpcV0_3_0Server for JsonRpcServerV0_3_0Impl {
         let mut old_root =
             GlobalRoot(StarkHash::try_from(GENESIS_HASH).map_err(internal_server_error)?);
         if parent_block_number.is_ok() {
-            let parent_header = get_block_header_by_number(&txn, parent_block_number.unwrap())?;
+            let parent_header: BlockHeader =
+                get_block_header_by_number(&txn, parent_block_number.unwrap())?;
             old_root = parent_header.new_root;
         }
 
@@ -203,7 +205,7 @@ impl JsonRpcV0_3_0Server for JsonRpcServerV0_3_0Impl {
 
         let block_number = transaction_index.0;
         let status = get_block_status(&txn, block_number)?;
-        let header =
+        let header: BlockHeader =
             get_block_header_by_number(&txn, block_number).map_err(internal_server_error)?;
 
         let transaction = txn
@@ -397,7 +399,7 @@ impl JsonRpcV0_3_0Server for JsonRpcServerV0_3_0Impl {
                         )?),
                     });
                 }
-                let header = get_block_header_by_number(&txn, block_number)
+                let header: BlockHeader = get_block_header_by_number(&txn, block_number)
                     .map_err(internal_server_error)?;
                 let transaction = txn
                     .get_transaction(event_index.0)
