+++ crates/papyrus_gateway/src/v0_3_0/block.rs
@@ -1,0 +1,36 @@
+use serde::{Deserialize, Serialize};
+use starknet_api::block::{BlockHash, BlockNumber, BlockStatus, BlockTimestamp};
+use starknet_api::core::{ContractAddress, GlobalRoot};
+
+use crate::transaction::Transactions;
+
+#[derive(Debug, Clone, Eq, PartialEq, Hash, Deserialize, Serialize, PartialOrd, Ord)]
+pub struct BlockHeader {
+    pub block_hash: BlockHash,
+    pub parent_hash: BlockHash,
+    pub block_number: BlockNumber,
+    pub sequencer_address: ContractAddress,
+    pub new_root: GlobalRoot,
+    pub timestamp: BlockTimestamp,
+}
+
+impl From<starknet_api::block::BlockHeader> for BlockHeader {
+    fn from(header: starknet_api::block::BlockHeader) -> Self {
+        BlockHeader {
+            block_hash: header.block_hash,
+            parent_hash: header.parent_hash,
+            block_number: header.block_number,
+            sequencer_address: header.sequencer,
+            new_root: header.state_root,
+            timestamp: header.timestamp,
+        }
+    }
+}
+
+#[derive(Debug, Clone, Eq, PartialEq, Hash, Deserialize, Serialize, PartialOrd, Ord)]
+pub struct Block {
+    pub status: BlockStatus,
+    #[serde(flatten)]
+    pub header: BlockHeader,
+    pub transactions: Transactions,
+}
