+++ crates/papyrus_gateway/src/v0_3_0/mod.rs
@@ -1,0 +1,1 @@
+pub mod block;
